rename `staderlabs_plugin.c` and `staderlabs_plugin.h`
