+++ src/handle_finalize.c
@@ -1,4 +1,4 @@
-#include "staderlabs_plugin.h"
+#include "plugin.h"
 
 void handle_finalize(ethPluginFinalize_t *msg) {
     context_t *context = (context_t *) msg->pluginContext;
