+++ src/handle_init_contract.c
@@ -1,5 +1,5 @@
 #include "plugin_utils.h"
-#include "staderlabs_plugin.h"
+#include "plugin.h"
 
 // Called once to init.
 void handle_init_contract(ethPluginInitContract_t *msg) {
