+++ src/handle_provide_parameter.c
@@ -1,4 +1,4 @@
-#include "staderlabs_plugin.h"
+#include "plugin.h"
 
 // Copy amount sent parameter to amount_received
 static void handle_amount_received(const ethPluginProvideParameter_t *msg, context_t *context) {
